Add runExtraTrainingAnalysis Training hook for spirit gauge analysis
